feat: introduce plugins infrastructure and add `OverlaysPlugin` (#62)
This infrastructure is a first try to propose extension points with
plugins. It is subject to change in the future.
The `OverlaysPlugin` demonstrates how to create and use a plugin. It
provides a method to show or hide overlays. They stay in the DOM
contrarily to the `BpmnElementsRegistry.removeOverlays` method that
actually removes overlays from the model.
Hiding overlays supports BPMN navigation which means that the overlays
stay hidden while panning and zooming.

Demo
  - create a dedicated demo for the `OverlaysPlugin`
- packaging: all pages of the demo are automatically packaged within the
distribution. Previously, they had to be declared manually.

Build of the 'addons' lib: generate sourceMap to easily debug the demo.