chore: add infrastructure to write JavaScript tests with jest (#100)
Jest is configured to run ESM tests.
This requires to patch bpmn-visualization as its package.json lacks of
ESM support in the latest available version. It will support it out of
the box in a future version (at which point we can remove the patch from
the project).

Writing tests in TypeScript will be possible at a later date.