docs: present the library as less "experimental" than before
We received several comments indicating that the term "experimental" might prevent people from
evaluating the library. They might see it as being at too early a stage of development.

We want to increase the dissemination and use of the library as much as possible.
That's why we've removed the word "experimental" from the wording.