feat(demo): improve zoom control buttons (#186)
Add title to buttons
Do not display "Fit" but use an icon instead. The related title is now
"Zoom reset" to use a more
common wording.

Zoom buttons are now managed by a dedicated component that hides the
button logic and the style.
This also paves the way for reuse of these buttons in other demos.