feat: introduce `StyleByNamePlugin`
Provide the same API as `StyleRegistry` but pass names instead of ids to identify BPMN elements.

NOTE: To map names to ids, the plugin uses `BpmnElementsSearcher`.
The mapping is currently not cached, nor pre-fetched after the BPMN source has been loaded.
So the implementation is not very effective. Caching and pre-fetch features will be implemented in
the future.