fix: add missing file extension in imports of the npm package (#301)
Webpack follows node.js rules for imports:
> In .mjs or .js files with a package.json whose ‘type’ is ‘module’,
imports must be fully specified.
This means that the extension must be included. This is how the ESM is
defined in node.js

The ‘type’ of the library is defined as ‘module’ in the package.json
file, so this behaviour is applied. So, the imports have been updated to
include the .js file extension. An eslint plugin is used to enforce and
automatically fix missing file extension in imports.

Previously, as the imports didn't include the file extension in version
0.7.0 of the library, webpack raised errors like in the following:

```
ERROR in ./node_modules/@process-analytics/bv-experimental-add-ons/dist/index.js 16:0-32
Module not found: Error: Can't resolve './bpmn-elements' in '/user/app/node_modules/@process-analytics/bv-experimental-add-ons/dist'
Did you mean 'bpmn-elements.js'?
BREAKING CHANGE: The request './bpmn-elements' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.

ERROR in ./node_modules/@process-analytics/bv-experimental-add-ons/dist/index.js 17:0-24
Module not found: Error: Can't resolve './paths' in '/user/app/node_modules/@process-analytics/bv-experimental-add-ons/dist'
Did you mean 'paths.js'?
BREAKING CHANGE: The request './paths' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.

ERROR in ./node_modules/@process-analytics/bv-experimental-add-ons/dist/index.js 18:0-26
Module not found: Error: Can't resolve './plugins' in '/user/app/node_modules/@process-analytics/bv-experimental-add-ons/dist'
Did you mean 'index.js'?
BREAKING CHANGE: The request './plugins' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.

ERROR in ./node_modules/@process-analytics/bv-experimental-add-ons/dist/index.js 19:0-34
Module not found: Error: Can't resolve './plugins-support' in '/user/app/node_modules/@process-analytics/bv-experimental-add-ons/dist'
Did you mean 'plugins-support.js'?
BREAKING CHANGE: The request './plugins-support' failed to resolve only because it was resolved as fully specified
...
```