chore(test): switch from ts-jest to @swc/jest (#302)
This makes the whole test run faster because `swc` uses native binaries
to transpile the TypeScript code.
This is particularly noticeable when you run a single test in your IDE.

The gain (i.e. the difference between execution with swc and with
ts-jest) will increase as there are more TypeScript files to transpile
(both in the tests and in the library).

Unlike `ts-jest`, `swc` does no type checking (it just transpiles). A
new npm script is available to type-check the test code.